Added mockery pass and worked it into the DI;
The mockery pass allows users to make a mockery object of any class that
the DI serves, this is particularly better way of doing it as it allows
the DI to be compiled, and also allows us to pass mockery classes around
inside the container. Both of which were not possible with the previous
mockery system.