Fixed mockery issue with DI by using a decorator class.
There were some issues with the current way I had used the Mockery
classes with the Symfony DI, it came down to the fact that Symfony
setClass expects a string and not an object, and so your mockery object
lost all its expectations during this. What I have done to Fix this is
add a Decorator class that acts as a proxy to the Mocked instances.