Fixtures. Doctrine Fixtures.
At the moment I have only test this with running basic fixtures, however
the Entity base class seems to have proved useful with this as it allows
us to control the entity within the fixture, without having the bridge a
connection through the fixture manager. This idea still needs expanding
however as its a very rough version of fixturing and probably not the
most efficient.